+++ .env
@@ -1,0 +1,1 @@
+EXPO_PUBLIC_STREAM_API_KEY=3sf3vhrwdmrt
