+++ app/(home)/_layout.tsx
@@ -1,39 +1,16 @@
-import { Slot, Stack } from "expo-router";
-import { useEffect } from "react";
-import { StreamChat } from "stream-chat";
-import { Chat, OverlayProvider } from "stream-chat-expo";
-
-const client = StreamChat.getInstance("3sf3vhrwdmrt");
+import ChatProvider from "@/providers/ChatProvider";
+import { Stack } from "expo-router";
 
 export default function HomeLayout() {
-
-    useEffect(() => {
-        const connect = async () => {
-            await client.connectUser(
-                {
-                    id: "jlahey",
-                    name: "Jim Lahey",
-                    image: "https://i.imgur.com/fR9Jz14.png",
-                },
-                client.devToken("jlahey"),
-            );
-
-        //     const channel = client.channel("messaging", "the_park", {
-        //         name: "The Park",
-        //     });
-        //    await channel.watch()
-        }
-        connect();
-    })
     return (
-        <OverlayProvider>
-            <Chat client={client}>
-               <Stack>
-                <Stack.Screen name="(tabs)" options={{headerShown:false}}/>
-                </Stack> 
-            </Chat>
-            
-        </OverlayProvider>
-        
+
+        <ChatProvider>
+            <Stack>
+                <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+            </Stack>
+        </ChatProvider>
+
+
+
     )
 }
